copy start_chaindata_updating from router
